Review remarks addressed
Changes to include materials sent in the in progress count depend
upon a change to the fusionApi script that are currently in a
branch awaiting review. Therefore, the rawgit URIs now refer to
the branch rather than master for the sake of testing.

When the fusionApi script is merged to its master,
we have to change the rawgit URIs.

Also sprinkled with comments.